Remove GPT and Google Vision from FormSync
