+++ back/Gemfile
@@ -167,7 +167,6 @@ gem 'intercom', '~> 4.2'
 gem 'mjml-rails', '~> 4.11'
 gem 'mrml'
 gem 'cloudfront-rails', '~> 0.4'
-gem 'google-cloud-document_ai', '~> 1.4'
 gem 'combine_pdf', '~> 1.0.26'
 gem 'neighbor', '~> 0.5.0'
 gem 'browser', '~> 6.1'
