+++ back/engines/commercial/bulk_import_ideas/app/services/bulk_import_ideas/parsers/pdf/idea_pdf_template_reader.rb
@@ -2,11 +2,10 @@
 
 module BulkImportIdeas::Parsers::Pdf
   class IdeaPdfTemplateReader
-    def initialize(phase, locale, personal_data_enabled, gpt_parser: false)
+    def initialize(phase, locale, personal_data_enabled, gpt_parser: false) # rubocop:disable Lint/UnusedMethodArgument
       @phase = phase
       @locale = locale
       @personal_data_enabled = personal_data_enabled
-      @gpt_parser = gpt_parser
     end
 
     # Extract the text from the template PDF so we understand how each field is laid out in the PDF
@@ -63,9 +62,9 @@ module BulkImportIdeas::Parsers::Pdf
 
       # Skip fields or options whose fields are not importable
       importable = if type == 'field'
-        @gpt_parser ? field_or_option.supports_pdf_gpt_import? : field_or_option.supports_pdf_import?
+        field_or_option.supports_pdf_import?
       else # option
-        @gpt_parser ? field_or_option.custom_field.supports_pdf_gpt_import? : field_or_option.custom_field.supports_pdf_import?
+        field_or_option.custom_field.supports_pdf_import?
       end
       return unless importable
