+++ back/engines/commercial/bulk_import_ideas/spec/acceptance/idea_import_spec.rb
@@ -334,9 +334,7 @@ resource 'BulkImportIdeasImportIdeas' do
   end
 
   def stub_external_apis
-    allow_any_instance_of(Analysis::LLM::GPT41).to receive(:chat).and_return('[{}]')
+    allow_any_instance_of(Analysis::LLM::ClaudeSonnet46).to receive(:chat).and_return('[{}]')
     allow_any_instance_of(BulkImportIdeas::Exporters::IdeaPdfFormExporter).to receive(:export).and_return(Rails.root.join('engines/commercial/bulk_import_ideas/spec/fixtures/scan_1.pdf'))
-    allow_any_instance_of(BulkImportIdeas::Parsers::Pdf::IdeaGoogleFormParserService).to receive(:raw_text_page_array).and_return(create_project_bulk_import_raw_text_array)
-    allow_any_instance_of(BulkImportIdeas::Parsers::Pdf::IdeaGoogleFormParserService).to receive(:parse_pdf).and_return(create_project_bulk_import_parse_pdf)
   end
 end
