+++ back/lib/input_type_strategy/file_upload.rb
@@ -2,7 +2,7 @@
 
 module InputTypeStrategy
   class FileUpload < Base
-    def supports_pdf_gpt_import?
+    def supports_pdf_import?
       false
     end
