+++ back/lib/input_type_strategy/image_files.rb
@@ -10,7 +10,7 @@ module InputTypeStrategy
       false
     end
 
-    def supports_pdf_gpt_import?
+    def supports_pdf_import?
       false
     end
