+++ back/lib/input_type_strategy/topic_ids.rb
@@ -6,7 +6,7 @@ module InputTypeStrategy
       true
     end
 
-    def supports_pdf_gpt_import?
+    def supports_pdf_import?
       false
     end
   end
