+++ front/app/containers/Admin/projects/project/inputImporter/ImportModal/ImportPdfModal.tsx
@@ -9,14 +9,12 @@ import {
 } from '@citizenlab/cl2-component-library';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { useForm, FormProvider } from 'react-hook-form';
-import { useParams, useSearchParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import { SupportedLocale } from 'typings';
 import { object, mixed, boolean } from 'yup';
 
 import { IBackgroundJobData } from 'api/background_jobs/types';
-import useAddOfflineIdeasAsync, {
-  ParserType,
-} from 'api/import_ideas/useAddOfflineIdeasAsync';
+import useAddOfflineIdeasAsync from 'api/import_ideas/useAddOfflineIdeasAsync';
 import usePhase from 'api/phases/usePhase';
 
 import useLocale from 'hooks/useLocale';
@@ -38,7 +36,6 @@ interface FormValues {
   locale: SupportedLocale;
   file: Record<string, any>;
   personal_data: boolean;
-  parser_consent: boolean;
 }
 
 interface Props {
@@ -57,11 +54,6 @@ const ImportPdfModal = ({ open, onClose, onImport }: Props) => {
     projectId: string;
   };
 
-  // Allows switching of different parsers if needed
-  const [searchParams] = useSearchParams();
-  const parser = (searchParams.get('parser') || undefined) as ParserType;
-  const parserConsentRequired = parser !== 'gpt' && parser !== 'claude'; // Only need to show consent when using Google document AI parser
-
   const downloadFormPath =
     phase?.data.attributes.participation_method === 'native_survey'
       ? `/admin/projects/${projectId}/phases/${phaseId}/survey-form`
@@ -71,23 +63,12 @@ const ImportPdfModal = ({ open, onClose, onImport }: Props) => {
     locale,
     file: undefined,
     personal_data: false,
-    parser_consent: !parserConsentRequired,
   };
 
   const schema = object({
     locale: validateLocale().required(),
     file: mixed().required(formatMessage(messages.pleaseUploadFile)),
-
     personal_data: boolean().required(),
-    parser_consent: boolean()
-      .required()
-      .test('', formatMessage(messages.consentNeeded), (v, context) => {
-        if (context.parent.file?.extension === 'application/pdf') {
-          return !!v;
-        }
-
-        return true;
-      }),
   });
 
   const methods = useForm<FormValues>({
@@ -96,12 +77,7 @@ const ImportPdfModal = ({ open, onClose, onImport }: Props) => {
     resolver: yupResolver(schema),
   });
 
-  const submitFile = async ({
-    file,
-    parser_consent: _,
-    locale,
-    personal_data,
-  }: FormValues) => {
+  const submitFile = async ({ file, locale, personal_data }: FormValues) => {
     if (!phaseId) return;
 
     try {
@@ -109,7 +85,6 @@ const ImportPdfModal = ({ open, onClose, onImport }: Props) => {
         phase_id: phaseId,
         file: file.base64,
         format: 'pdf',
-        parser,
         locale,
         personal_data,
       });
@@ -169,14 +144,6 @@ const ImportPdfModal = ({ open, onClose, onImport }: Props) => {
                 label={<FormattedMessage {...messages.formHasPersonalData} />}
               />
             </Box>
-            {parserConsentRequired && (
-              <Box mt="24px">
-                <CheckboxWithLabel
-                  name="parser_consent"
-                  label={<FormattedMessage {...messages.googleConsent} />}
-                />
-              </Box>
-            )}
             <Box w="100%" display="flex" mt="32px">
               <Button
                 bgColor={colors.primary}
